PP-4516 Handle error http codes from downstream services
Previously when call ing a downstream service we counted
anything that wasn't an error as a  successful call to a
service (either to connector or to
perf platform).
The `request` library only sets the `error`
object if the http call has failed - e.g the connection timed
out or wasn't a connection was refused - it doesn't include
HTTP error codes. This commit checks that the call recieves a 200
(a bit simplistic, but correct for this case) and if not
rejects the promise, ensuring we log something useful and don't
believe all is fine when it patently is not.